requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
streamlit==1.35.0
streamlit-folium==0.20.0
folium==0.16.0
scikit-learn==1.4.2
pandas==2.2.2
geopandas==0.14.4
numpy==1.26.4
shapely==2.0.4
requests==2.32.2
python-dotenv==1.0.1
openai==1.30.1
httpx==0.27.0
